backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.4.2
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1